Extract information from fitted EC50 objects

These helpers make fitted objects from [estimate_EC50()] and [ec50_multimodel()] easier to reuse. They return plain data frames, metadata, stored `drc` models, or fitted curve coordinates without requiring users to work with object attributes directly.

Usage

ec50_estimates(x)

ec50_metadata(x)

fitted_models(x)

curve_data(x, n_points = 200, log_x = TRUE, quiet = FALSE)

Arguments

x

An object returned by [estimate_EC50()] or [ec50_multimodel()].

n_points

Number of dose values used to draw each fitted curve.

log_x

Logical. If `TRUE`, curve coordinates are generated only over positive dose values for log10 x-axis plotting.

quiet

Logical. If `FALSE`, failed curve predictions are reported with a warning.

Value

`ec50_estimates()` returns a plain data frame of EC estimates. `ec50_metadata()` returns a list with modeling metadata. `fitted_models()` returns a named list of stored `drc` model objects. `curve_data()` returns a data frame of fitted curve coordinates.

Examples

data(multi_isolate)
sample_data <- subset(
  multi_isolate,
  isolate %in% 1:3 & fungicida == "Fungicide A"
)

fit <- estimate_EC50(
  growth ~ dose,
  data = sample_data,
  isolate_col = "isolate",
  strata_col = "field",
  fct = drc::LL.3()
)

ec50_estimates(fit)
#>   ID        field    Estimate   Std..Error
#> 1  1      Organic 0.006072082 0.0005740341
#> 2  2 Conventional 0.101455765 0.0076364691
#> 3  3      Organic 0.003776957 0.0002432571
ec50_metadata(fit)
#> $formula
#> growth ~ dose
#> <environment: 0x000001ffc3d082e0>
#> 
#> $data_columns
#> [1] "isolate"   "field"     "fungicida" "dose"      "growth"   
#> 
#> $isolate_col
#> [1] "isolate"
#> 
#> $strata_col
#> [1] "field"
#> 
#> $model_labels
#> [1] "LL.3"
#> 
#> $n_models
#> [1] 3
#> 
curve_data(fit)
#>               field isolate model         dose      growth        .curve_group
#> 1           Organic       1  LL.3 1.000000e-05 19.86338289      Organic.1.LL.3
#> 1.1         Organic       1  LL.3 1.059560e-05 19.86006320      Organic.1.LL.3
#> 1.2         Organic       1  LL.3 1.122668e-05 19.85656741      Organic.1.LL.3
#> 1.3         Organic       1  LL.3 1.189534e-05 19.85288626      Organic.1.LL.3
#> 1.4         Organic       1  LL.3 1.260383e-05 19.84900999      Organic.1.LL.3
#> 1.5         Organic       1  LL.3 1.335452e-05 19.84492834      Organic.1.LL.3
#> 1.6         Organic       1  LL.3 1.414991e-05 19.84063051      Organic.1.LL.3
#> 1.7         Organic       1  LL.3 1.499268e-05 19.83610516      Organic.1.LL.3
#> 1.8         Organic       1  LL.3 1.588565e-05 19.83134037      Organic.1.LL.3
#> 1.9         Organic       1  LL.3 1.683180e-05 19.82632358      Organic.1.LL.3
#> 1.10        Organic       1  LL.3 1.783431e-05 19.82104160      Organic.1.LL.3
#> 1.11        Organic       1  LL.3 1.889652e-05 19.81548057      Organic.1.LL.3
#> 1.12        Organic       1  LL.3 2.002200e-05 19.80962592      Organic.1.LL.3
#> 1.13        Organic       1  LL.3 2.121452e-05 19.80346234      Organic.1.LL.3
#> 1.14        Organic       1  LL.3 2.247806e-05 19.79697372      Organic.1.LL.3
#> 1.15        Organic       1  LL.3 2.381686e-05 19.79014317      Organic.1.LL.3
#> 1.16        Organic       1  LL.3 2.523539e-05 19.78295292      Organic.1.LL.3
#> 1.17        Organic       1  LL.3 2.673842e-05 19.77538430      Organic.1.LL.3
#> 1.18        Organic       1  LL.3 2.833096e-05 19.76741774      Organic.1.LL.3
#> 1.19        Organic       1  LL.3 3.001836e-05 19.75903265      Organic.1.LL.3
#> 1.20        Organic       1  LL.3 3.180626e-05 19.75020743      Organic.1.LL.3
#> 1.21        Organic       1  LL.3 3.370064e-05 19.74091941      Organic.1.LL.3
#> 1.22        Organic       1  LL.3 3.570786e-05 19.73114479      Organic.1.LL.3
#> 1.23        Organic       1  LL.3 3.783463e-05 19.72085860      Organic.1.LL.3
#> 1.24        Organic       1  LL.3 4.008806e-05 19.71003466      Organic.1.LL.3
#> 1.25        Organic       1  LL.3 4.247572e-05 19.69864548      Organic.1.LL.3
#> 1.26        Organic       1  LL.3 4.500558e-05 19.68666228      Organic.1.LL.3
#> 1.27        Organic       1  LL.3 4.768612e-05 19.67405486      Organic.1.LL.3
#> 1.28        Organic       1  LL.3 5.052631e-05 19.66079157      Organic.1.LL.3
#> 1.29        Organic       1  LL.3 5.353567e-05 19.64683926      Organic.1.LL.3
#> 1.30        Organic       1  LL.3 5.672426e-05 19.63216320      Organic.1.LL.3
#> 1.31        Organic       1  LL.3 6.010277e-05 19.61672704      Organic.1.LL.3
#> 1.32        Organic       1  LL.3 6.368250e-05 19.60049272      Organic.1.LL.3
#> 1.33        Organic       1  LL.3 6.747544e-05 19.58342042      Organic.1.LL.3
#> 1.34        Organic       1  LL.3 7.149429e-05 19.56546848      Organic.1.LL.3
#> 1.35        Organic       1  LL.3 7.575250e-05 19.54659335      Organic.1.LL.3
#> 1.36        Organic       1  LL.3 8.026434e-05 19.52674951      Organic.1.LL.3
#> 1.37        Organic       1  LL.3 8.504489e-05 19.50588941      Organic.1.LL.3
#> 1.38        Organic       1  LL.3 9.011018e-05 19.48396339      Organic.1.LL.3
#> 1.39        Organic       1  LL.3 9.547716e-05 19.46091963      Organic.1.LL.3
#> 1.40        Organic       1  LL.3 1.011638e-04 19.43670406      Organic.1.LL.3
#> 1.41        Organic       1  LL.3 1.071891e-04 19.41126031      Organic.1.LL.3
#> 1.42        Organic       1  LL.3 1.135733e-04 19.38452964      Organic.1.LL.3
#> 1.43        Organic       1  LL.3 1.203378e-04 19.35645088      Organic.1.LL.3
#> 1.44        Organic       1  LL.3 1.275051e-04 19.32696038      Organic.1.LL.3
#> 1.45        Organic       1  LL.3 1.350994e-04 19.29599192      Organic.1.LL.3
#> 1.46        Organic       1  LL.3 1.431459e-04 19.26347670      Organic.1.LL.3
#> 1.47        Organic       1  LL.3 1.516717e-04 19.22934326      Organic.1.LL.3
#> 1.48        Organic       1  LL.3 1.607053e-04 19.19351745      Organic.1.LL.3
#> 1.49        Organic       1  LL.3 1.702769e-04 19.15592240      Organic.1.LL.3
#> 1.50        Organic       1  LL.3 1.804186e-04 19.11647849      Organic.1.LL.3
#> 1.51        Organic       1  LL.3 1.911644e-04 19.07510331      Organic.1.LL.3
#> 1.52        Organic       1  LL.3 2.025502e-04 19.03171170      Organic.1.LL.3
#> 1.53        Organic       1  LL.3 2.146141e-04 18.98621568      Organic.1.LL.3
#> 1.54        Organic       1  LL.3 2.273966e-04 18.93852454      Organic.1.LL.3
#> 1.55        Organic       1  LL.3 2.409404e-04 18.88854482      Organic.1.LL.3
#> 1.56        Organic       1  LL.3 2.552908e-04 18.83618037      Organic.1.LL.3
#> 1.57        Organic       1  LL.3 2.704960e-04 18.78133241      Organic.1.LL.3
#> 1.58        Organic       1  LL.3 2.866068e-04 18.72389961      Organic.1.LL.3
#> 1.59        Organic       1  LL.3 3.036771e-04 18.66377820      Organic.1.LL.3
#> 1.60        Organic       1  LL.3 3.217642e-04 18.60086208      Organic.1.LL.3
#> 1.61        Organic       1  LL.3 3.409285e-04 18.53504297      Organic.1.LL.3
#> 1.62        Organic       1  LL.3 3.612343e-04 18.46621059      Organic.1.LL.3
#> 1.63        Organic       1  LL.3 3.827494e-04 18.39425288      Organic.1.LL.3
#> 1.64        Organic       1  LL.3 4.055461e-04 18.31905620      Organic.1.LL.3
#> 1.65        Organic       1  LL.3 4.297005e-04 18.24050562      Organic.1.LL.3
#> 1.66        Organic       1  LL.3 4.552935e-04 18.15848520      Organic.1.LL.3
#> 1.67        Organic       1  LL.3 4.824109e-04 18.07287837      Organic.1.LL.3
#> 1.68        Organic       1  LL.3 5.111433e-04 17.98356823      Organic.1.LL.3
#> 1.69        Organic       1  LL.3 5.415871e-04 17.89043806      Organic.1.LL.3
#> 1.70        Organic       1  LL.3 5.738442e-04 17.79337169      Organic.1.LL.3
#> 1.71        Organic       1  LL.3 6.080224e-04 17.69225402      Organic.1.LL.3
#> 1.72        Organic       1  LL.3 6.442364e-04 17.58697160      Organic.1.LL.3
#> 1.73        Organic       1  LL.3 6.826072e-04 17.47741314      Organic.1.LL.3
#> 1.74        Organic       1  LL.3 7.232634e-04 17.36347021      Organic.1.LL.3
#> 1.75        Organic       1  LL.3 7.663411e-04 17.24503785      Organic.1.LL.3
#> 1.76        Organic       1  LL.3 8.119845e-04 17.12201528      Organic.1.LL.3
#> 1.77        Organic       1  LL.3 8.603464e-04 16.99430669      Organic.1.LL.3
#> 1.78        Organic       1  LL.3 9.115888e-04 16.86182198      Organic.1.LL.3
#> 1.79        Organic       1  LL.3 9.658832e-04 16.72447760      Organic.1.LL.3
#> 1.80        Organic       1  LL.3 1.023411e-03 16.58219738      Organic.1.LL.3
#> 1.81        Organic       1  LL.3 1.084366e-03 16.43491344      Organic.1.LL.3
#> 1.82        Organic       1  LL.3 1.148951e-03 16.28256703      Organic.1.LL.3
#> 1.83        Organic       1  LL.3 1.217383e-03 16.12510948      Organic.1.LL.3
#> 1.84        Organic       1  LL.3 1.289890e-03 15.96250308      Organic.1.LL.3
#> 1.85        Organic       1  LL.3 1.366716e-03 15.79472200      Organic.1.LL.3
#> 1.86        Organic       1  LL.3 1.448118e-03 15.62175319      Organic.1.LL.3
#> 1.87        Organic       1  LL.3 1.534368e-03 15.44359726      Organic.1.LL.3
#> 1.88        Organic       1  LL.3 1.625756e-03 15.26026928      Organic.1.LL.3
#> 1.89        Organic       1  LL.3 1.722586e-03 15.07179963      Organic.1.LL.3
#> 1.90        Organic       1  LL.3 1.825183e-03 14.87823470      Organic.1.LL.3
#> 1.91        Organic       1  LL.3 1.933892e-03 14.67963760      Organic.1.LL.3
#> 1.92        Organic       1  LL.3 2.049075e-03 14.47608870      Organic.1.LL.3
#> 1.93        Organic       1  LL.3 2.171118e-03 14.26768616      Organic.1.LL.3
#> 1.94        Organic       1  LL.3 2.300430e-03 14.05454632      Organic.1.LL.3
#> 1.95        Organic       1  LL.3 2.437444e-03 13.83680397      Organic.1.LL.3
#> 1.96        Organic       1  LL.3 2.582619e-03 13.61461250      Organic.1.LL.3
#> 1.97        Organic       1  LL.3 2.736440e-03 13.38814391      Organic.1.LL.3
#> 1.98        Organic       1  LL.3 2.899423e-03 13.15758864      Organic.1.LL.3
#> 1.99        Organic       1  LL.3 3.072113e-03 12.92315537      Organic.1.LL.3
#> 1.100       Organic       1  LL.3 3.255089e-03 12.68507046      Organic.1.LL.3
#> 1.101       Organic       1  LL.3 3.448962e-03 12.44357744      Organic.1.LL.3
#> 1.102       Organic       1  LL.3 3.654383e-03 12.19893619      Organic.1.LL.3
#> 1.103       Organic       1  LL.3 3.872039e-03 11.95142203      Organic.1.LL.3
#> 1.104       Organic       1  LL.3 4.102658e-03 11.70132465      Organic.1.LL.3
#> 1.105       Organic       1  LL.3 4.347013e-03 11.44894687      Organic.1.LL.3
#> 1.106       Organic       1  LL.3 4.605922e-03 11.19460326      Organic.1.LL.3
#> 1.107       Organic       1  LL.3 4.880252e-03 10.93861868      Organic.1.LL.3
#> 1.108       Organic       1  LL.3 5.170920e-03 10.68132666      Organic.1.LL.3
#> 1.109       Organic       1  LL.3 5.478901e-03 10.42306769      Organic.1.LL.3
#> 1.110       Organic       1  LL.3 5.805226e-03 10.16418744      Organic.1.LL.3
#> 1.111       Organic       1  LL.3 6.150986e-03  9.90503494      Organic.1.LL.3
#> 1.112       Organic       1  LL.3 6.517340e-03  9.64596069      Organic.1.LL.3
#> 1.113       Organic       1  LL.3 6.905514e-03  9.38731477      Organic.1.LL.3
#> 1.114       Organic       1  LL.3 7.316807e-03  9.12944492      Organic.1.LL.3
#> 1.115       Organic       1  LL.3 7.752597e-03  8.87269475      Organic.1.LL.3
#> 1.116       Organic       1  LL.3 8.214344e-03  8.61740185      Organic.1.LL.3
#> 1.117       Organic       1  LL.3 8.703591e-03  8.36389607      Organic.1.LL.3
#> 1.118       Organic       1  LL.3 9.221979e-03  8.11249788      Organic.1.LL.3
#> 1.119       Organic       1  LL.3 9.771242e-03  7.86351680      Organic.1.LL.3
#> 1.120       Organic       1  LL.3 1.035322e-02  7.61724993      Organic.1.LL.3
#> 1.121       Organic       1  LL.3 1.096986e-02  7.37398069      Organic.1.LL.3
#> 1.122       Organic       1  LL.3 1.162322e-02  7.13397764      Organic.1.LL.3
#> 1.123       Organic       1  LL.3 1.231551e-02  6.89749346      Organic.1.LL.3
#> 1.124       Organic       1  LL.3 1.304902e-02  6.66476411      Organic.1.LL.3
#> 1.125       Organic       1  LL.3 1.382622e-02  6.43600812      Organic.1.LL.3
#> 1.126       Organic       1  LL.3 1.464971e-02  6.21142607      Organic.1.LL.3
#> 1.127       Organic       1  LL.3 1.552225e-02  5.99120021      Organic.1.LL.3
#> 1.128       Organic       1  LL.3 1.644676e-02  5.77549423      Organic.1.LL.3
#> 1.129       Organic       1  LL.3 1.742633e-02  5.56445322      Organic.1.LL.3
#> 1.130       Organic       1  LL.3 1.846425e-02  5.35820369      Organic.1.LL.3
#> 1.131       Organic       1  LL.3 1.956398e-02  5.15685384      Organic.1.LL.3
#> 1.132       Organic       1  LL.3 2.072922e-02  4.96049384      Organic.1.LL.3
#> 1.133       Organic       1  LL.3 2.196385e-02  4.76919627      Organic.1.LL.3
#> 1.134       Organic       1  LL.3 2.327202e-02  4.58301672      Organic.1.LL.3
#> 1.135       Organic       1  LL.3 2.465811e-02  4.40199432      Organic.1.LL.3
#> 1.136       Organic       1  LL.3 2.612675e-02  4.22615255      Organic.1.LL.3
#> 1.137       Organic       1  LL.3 2.768287e-02  4.05549991      Organic.1.LL.3
#> 1.138       Organic       1  LL.3 2.933166e-02  3.89003081      Organic.1.LL.3
#> 1.139       Organic       1  LL.3 3.107866e-02  3.72972639      Organic.1.LL.3
#> 1.140       Organic       1  LL.3 3.292971e-02  3.57455540      Organic.1.LL.3
#> 1.141       Organic       1  LL.3 3.489101e-02  3.42447513      Organic.1.LL.3
#> 1.142       Organic       1  LL.3 3.696913e-02  3.27943230      Organic.1.LL.3
#> 1.143       Organic       1  LL.3 3.917101e-02  3.13936398      Organic.1.LL.3
#> 1.144       Organic       1  LL.3 4.150405e-02  3.00419846      Organic.1.LL.3
#> 1.145       Organic       1  LL.3 4.397604e-02  2.87385618      Organic.1.LL.3
#> 1.146       Organic       1  LL.3 4.659526e-02  2.74825057      Organic.1.LL.3
#> 1.147       Organic       1  LL.3 4.937048e-02  2.62728887      Organic.1.LL.3
#> 1.148       Organic       1  LL.3 5.231099e-02  2.51087297      Organic.1.LL.3
#> 1.149       Organic       1  LL.3 5.542665e-02  2.39890013      Organic.1.LL.3
#> 1.150       Organic       1  LL.3 5.872787e-02  2.29126376      Organic.1.LL.3
#> 1.151       Organic       1  LL.3 6.222571e-02  2.18785407      Organic.1.LL.3
#> 1.152       Organic       1  LL.3 6.593188e-02  2.08855872      Organic.1.LL.3
#> 1.153       Organic       1  LL.3 6.985880e-02  1.99326347      Organic.1.LL.3
#> 1.154       Organic       1  LL.3 7.401960e-02  1.90185267      Organic.1.LL.3
#> 1.155       Organic       1  LL.3 7.842822e-02  1.81420986      Organic.1.LL.3
#> 1.156       Organic       1  LL.3 8.309942e-02  1.73021816      Organic.1.LL.3
#> 1.157       Organic       1  LL.3 8.804884e-02  1.64976079      Organic.1.LL.3
#> 1.158       Organic       1  LL.3 9.329304e-02  1.57272141      Organic.1.LL.3
#> 1.159       Organic       1  LL.3 9.884959e-02  1.49898449      Organic.1.LL.3
#> 1.160       Organic       1  LL.3 1.047371e-01  1.42843565      Organic.1.LL.3
#> 1.161       Organic       1  LL.3 1.109752e-01  1.36096191      Organic.1.LL.3
#> 1.162       Organic       1  LL.3 1.175850e-01  1.29645199      Organic.1.LL.3
#> 1.163       Organic       1  LL.3 1.245883e-01  1.23479647      Organic.1.LL.3
#> 1.164       Organic       1  LL.3 1.320088e-01  1.17588803      Organic.1.LL.3
#> 1.165       Organic       1  LL.3 1.398713e-01  1.11962159      Organic.1.LL.3
#> 1.166       Organic       1  LL.3 1.482021e-01  1.06589445      Organic.1.LL.3
#> 1.167       Organic       1  LL.3 1.570290e-01  1.01460641      Organic.1.LL.3
#> 1.168       Organic       1  LL.3 1.663817e-01  0.96565986      Organic.1.LL.3
#> 1.169       Organic       1  LL.3 1.762914e-01  0.91895985      Organic.1.LL.3
#> 1.170       Organic       1  LL.3 1.867914e-01  0.87441412      Organic.1.LL.3
#> 1.171       Organic       1  LL.3 1.979167e-01  0.83193319      Organic.1.LL.3
#> 1.172       Organic       1  LL.3 2.097046e-01  0.79143035      Organic.1.LL.3
#> 1.173       Organic       1  LL.3 2.221947e-01  0.75282164      Organic.1.LL.3
#> 1.174       Organic       1  LL.3 2.354286e-01  0.71602592      Organic.1.LL.3
#> 1.175       Organic       1  LL.3 2.494508e-01  0.68096479      Organic.1.LL.3
#> 1.176       Organic       1  LL.3 2.643081e-01  0.64756261      Organic.1.LL.3
#> 1.177       Organic       1  LL.3 2.800504e-01  0.61574641      Organic.1.LL.3
#> 1.178       Organic       1  LL.3 2.967302e-01  0.58544595      Organic.1.LL.3
#> 1.179       Organic       1  LL.3 3.144035e-01  0.55659357      Organic.1.LL.3
#> 1.180       Organic       1  LL.3 3.331295e-01  0.52912422      Organic.1.LL.3
#> 1.181       Organic       1  LL.3 3.529707e-01  0.50297535      Organic.1.LL.3
#> 1.182       Organic       1  LL.3 3.739937e-01  0.47808688      Organic.1.LL.3
#> 1.183       Organic       1  LL.3 3.962689e-01  0.45440114      Organic.1.LL.3
#> 1.184       Organic       1  LL.3 4.198707e-01  0.43186280      Organic.1.LL.3
#> 1.185       Organic       1  LL.3 4.448783e-01  0.41041880      Organic.1.LL.3
#> 1.186       Organic       1  LL.3 4.713753e-01  0.39001829      Organic.1.LL.3
#> 1.187       Organic       1  LL.3 4.994505e-01  0.37061256      Organic.1.LL.3
#> 1.188       Organic       1  LL.3 5.291979e-01  0.35215498      Organic.1.LL.3
#> 1.189       Organic       1  LL.3 5.607170e-01  0.33460091      Organic.1.LL.3
#> 1.190       Organic       1  LL.3 5.941134e-01  0.31790766      Organic.1.LL.3
#> 1.191       Organic       1  LL.3 6.294989e-01  0.30203439      Organic.1.LL.3
#> 1.192       Organic       1  LL.3 6.669920e-01  0.28694208      Organic.1.LL.3
#> 1.193       Organic       1  LL.3 7.067181e-01  0.27259345      Organic.1.LL.3
#> 1.194       Organic       1  LL.3 7.488104e-01  0.25895286      Organic.1.LL.3
#> 1.195       Organic       1  LL.3 7.934097e-01  0.24598630      Organic.1.LL.3
#> 1.196       Organic       1  LL.3 8.406653e-01  0.23366130      Organic.1.LL.3
#> 1.197       Organic       1  LL.3 8.907355e-01  0.22194688      Organic.1.LL.3
#> 1.198       Organic       1  LL.3 9.437878e-01  0.21081346      Organic.1.LL.3
#> 1.199       Organic       1  LL.3 1.000000e+00  0.20023285      Organic.1.LL.3
#> 36     Conventional       2  LL.3 1.000000e-05 20.00169656 Conventional.2.LL.3
#> 36.1   Conventional       2  LL.3 1.059560e-05 20.00159567 Conventional.2.LL.3
#> 36.2   Conventional       2  LL.3 1.122668e-05 20.00148866 Conventional.2.LL.3
#> 36.3   Conventional       2  LL.3 1.189534e-05 20.00137515 Conventional.2.LL.3
#> 36.4   Conventional       2  LL.3 1.260383e-05 20.00125476 Conventional.2.LL.3
#> 36.5   Conventional       2  LL.3 1.335452e-05 20.00112706 Conventional.2.LL.3
#> 36.6   Conventional       2  LL.3 1.414991e-05 20.00099162 Conventional.2.LL.3
#> 36.7   Conventional       2  LL.3 1.499268e-05 20.00084795 Conventional.2.LL.3
#> 36.8   Conventional       2  LL.3 1.588565e-05 20.00069557 Conventional.2.LL.3
#> 36.9   Conventional       2  LL.3 1.683180e-05 20.00053394 Conventional.2.LL.3
#> 36.10  Conventional       2  LL.3 1.783431e-05 20.00036251 Conventional.2.LL.3
#> 36.11  Conventional       2  LL.3 1.889652e-05 20.00018067 Conventional.2.LL.3
#> 36.12  Conventional       2  LL.3 2.002200e-05 19.99998780 Conventional.2.LL.3
#> 36.13  Conventional       2  LL.3 2.121452e-05 19.99978322 Conventional.2.LL.3
#> 36.14  Conventional       2  LL.3 2.247806e-05 19.99956624 Conventional.2.LL.3
#> 36.15  Conventional       2  LL.3 2.381686e-05 19.99933609 Conventional.2.LL.3
#> 36.16  Conventional       2  LL.3 2.523539e-05 19.99909198 Conventional.2.LL.3
#> 36.17  Conventional       2  LL.3 2.673842e-05 19.99883306 Conventional.2.LL.3
#> 36.18  Conventional       2  LL.3 2.833096e-05 19.99855843 Conventional.2.LL.3
#> 36.19  Conventional       2  LL.3 3.001836e-05 19.99826714 Conventional.2.LL.3
#> 36.20  Conventional       2  LL.3 3.180626e-05 19.99795818 Conventional.2.LL.3
#> 36.21  Conventional       2  LL.3 3.370064e-05 19.99763048 Conventional.2.LL.3
#> 36.22  Conventional       2  LL.3 3.570786e-05 19.99728290 Conventional.2.LL.3
#> 36.23  Conventional       2  LL.3 3.783463e-05 19.99691424 Conventional.2.LL.3
#> 36.24  Conventional       2  LL.3 4.008806e-05 19.99652322 Conventional.2.LL.3
#> 36.25  Conventional       2  LL.3 4.247572e-05 19.99610848 Conventional.2.LL.3
#> 36.26  Conventional       2  LL.3 4.500558e-05 19.99566859 Conventional.2.LL.3
#> 36.27  Conventional       2  LL.3 4.768612e-05 19.99520202 Conventional.2.LL.3
#> 36.28  Conventional       2  LL.3 5.052631e-05 19.99470716 Conventional.2.LL.3
#> 36.29  Conventional       2  LL.3 5.353567e-05 19.99418229 Conventional.2.LL.3
#> 36.30  Conventional       2  LL.3 5.672426e-05 19.99362559 Conventional.2.LL.3
#> 36.31  Conventional       2  LL.3 6.010277e-05 19.99303514 Conventional.2.LL.3
#> 36.32  Conventional       2  LL.3 6.368250e-05 19.99240889 Conventional.2.LL.3
#> 36.33  Conventional       2  LL.3 6.747544e-05 19.99174468 Conventional.2.LL.3
#> 36.34  Conventional       2  LL.3 7.149429e-05 19.99104019 Conventional.2.LL.3
#> 36.35  Conventional       2  LL.3 7.575250e-05 19.99029301 Conventional.2.LL.3
#> 36.36  Conventional       2  LL.3 8.026434e-05 19.98950054 Conventional.2.LL.3
#> 36.37  Conventional       2  LL.3 8.504489e-05 19.98866003 Conventional.2.LL.3
#> 36.38  Conventional       2  LL.3 9.011018e-05 19.98776859 Conventional.2.LL.3
#> 36.39  Conventional       2  LL.3 9.547716e-05 19.98682313 Conventional.2.LL.3
#> 36.40  Conventional       2  LL.3 1.011638e-04 19.98582039 Conventional.2.LL.3
#> 36.41  Conventional       2  LL.3 1.071891e-04 19.98475690 Conventional.2.LL.3
#> 36.42  Conventional       2  LL.3 1.135733e-04 19.98362898 Conventional.2.LL.3
#> 36.43  Conventional       2  LL.3 1.203378e-04 19.98243274 Conventional.2.LL.3
#> 36.44  Conventional       2  LL.3 1.275051e-04 19.98116406 Conventional.2.LL.3
#> 36.45  Conventional       2  LL.3 1.350994e-04 19.97981855 Conventional.2.LL.3
#> 36.46  Conventional       2  LL.3 1.431459e-04 19.97839157 Conventional.2.LL.3
#> 36.47  Conventional       2  LL.3 1.516717e-04 19.97687821 Conventional.2.LL.3
#> 36.48  Conventional       2  LL.3 1.607053e-04 19.97527325 Conventional.2.LL.3
#> 36.49  Conventional       2  LL.3 1.702769e-04 19.97357116 Conventional.2.LL.3
#> 36.50  Conventional       2  LL.3 1.804186e-04 19.97176608 Conventional.2.LL.3
#> 36.51  Conventional       2  LL.3 1.911644e-04 19.96985179 Conventional.2.LL.3
#> 36.52  Conventional       2  LL.3 2.025502e-04 19.96782173 Conventional.2.LL.3
#> 36.53  Conventional       2  LL.3 2.146141e-04 19.96566890 Conventional.2.LL.3
#> 36.54  Conventional       2  LL.3 2.273966e-04 19.96338591 Conventional.2.LL.3
#> 36.55  Conventional       2  LL.3 2.409404e-04 19.96096494 Conventional.2.LL.3
#> 36.56  Conventional       2  LL.3 2.552908e-04 19.95839766 Conventional.2.LL.3
#> 36.57  Conventional       2  LL.3 2.704960e-04 19.95567529 Conventional.2.LL.3
#> 36.58  Conventional       2  LL.3 2.866068e-04 19.95278850 Conventional.2.LL.3
#> 36.59  Conventional       2  LL.3 3.036771e-04 19.94972741 Conventional.2.LL.3
#> 36.60  Conventional       2  LL.3 3.217642e-04 19.94648156 Conventional.2.LL.3
#> 36.61  Conventional       2  LL.3 3.409285e-04 19.94303986 Conventional.2.LL.3
#> 36.62  Conventional       2  LL.3 3.612343e-04 19.93939057 Conventional.2.LL.3
#> 36.63  Conventional       2  LL.3 3.827494e-04 19.93552125 Conventional.2.LL.3
#> 36.64  Conventional       2  LL.3 4.055461e-04 19.93141872 Conventional.2.LL.3
#> 36.65  Conventional       2  LL.3 4.297005e-04 19.92706904 Conventional.2.LL.3
#> 36.66  Conventional       2  LL.3 4.552935e-04 19.92245744 Conventional.2.LL.3
#> 36.67  Conventional       2  LL.3 4.824109e-04 19.91756827 Conventional.2.LL.3
#> 36.68  Conventional       2  LL.3 5.111433e-04 19.91238498 Conventional.2.LL.3
#> 36.69  Conventional       2  LL.3 5.415871e-04 19.90689005 Conventional.2.LL.3
#> 36.70  Conventional       2  LL.3 5.738442e-04 19.90106492 Conventional.2.LL.3
#> 36.71  Conventional       2  LL.3 6.080224e-04 19.89488996 Conventional.2.LL.3
#> 36.72  Conventional       2  LL.3 6.442364e-04 19.88834441 Conventional.2.LL.3
#> 36.73  Conventional       2  LL.3 6.826072e-04 19.88140629 Conventional.2.LL.3
#> 36.74  Conventional       2  LL.3 7.232634e-04 19.87405236 Conventional.2.LL.3
#> 36.75  Conventional       2  LL.3 7.663411e-04 19.86625803 Conventional.2.LL.3
#> 36.76  Conventional       2  LL.3 8.119845e-04 19.85799732 Conventional.2.LL.3
#> 36.77  Conventional       2  LL.3 8.603464e-04 19.84924273 Conventional.2.LL.3
#> 36.78  Conventional       2  LL.3 9.115888e-04 19.83996522 Conventional.2.LL.3
#> 36.79  Conventional       2  LL.3 9.658832e-04 19.83013411 Conventional.2.LL.3
#> 36.80  Conventional       2  LL.3 1.023411e-03 19.81971694 Conventional.2.LL.3
#> 36.81  Conventional       2  LL.3 1.084366e-03 19.80867947 Conventional.2.LL.3
#> 36.82  Conventional       2  LL.3 1.148951e-03 19.79698552 Conventional.2.LL.3
#> 36.83  Conventional       2  LL.3 1.217383e-03 19.78459690 Conventional.2.LL.3
#> 36.84  Conventional       2  LL.3 1.289890e-03 19.77147331 Conventional.2.LL.3
#> 36.85  Conventional       2  LL.3 1.366716e-03 19.75757220 Conventional.2.LL.3
#> 36.86  Conventional       2  LL.3 1.448118e-03 19.74284873 Conventional.2.LL.3
#> 36.87  Conventional       2  LL.3 1.534368e-03 19.72725560 Conventional.2.LL.3
#> 36.88  Conventional       2  LL.3 1.625756e-03 19.71074295 Conventional.2.LL.3
#> 36.89  Conventional       2  LL.3 1.722586e-03 19.69325828 Conventional.2.LL.3
#> 36.90  Conventional       2  LL.3 1.825183e-03 19.67474626 Conventional.2.LL.3
#> 36.91  Conventional       2  LL.3 1.933892e-03 19.65514868 Conventional.2.LL.3
#> 36.92  Conventional       2  LL.3 2.049075e-03 19.63440426 Conventional.2.LL.3
#> 36.93  Conventional       2  LL.3 2.171118e-03 19.61244859 Conventional.2.LL.3
#> 36.94  Conventional       2  LL.3 2.300430e-03 19.58921394 Conventional.2.LL.3
#> 36.95  Conventional       2  LL.3 2.437444e-03 19.56462918 Conventional.2.LL.3
#> 36.96  Conventional       2  LL.3 2.582619e-03 19.53861963 Conventional.2.LL.3
#> 36.97  Conventional       2  LL.3 2.736440e-03 19.51110692 Conventional.2.LL.3
#> 36.98  Conventional       2  LL.3 2.899423e-03 19.48200892 Conventional.2.LL.3
#> 36.99  Conventional       2  LL.3 3.072113e-03 19.45123955 Conventional.2.LL.3
#> 36.100 Conventional       2  LL.3 3.255089e-03 19.41870872 Conventional.2.LL.3
#> 36.101 Conventional       2  LL.3 3.448962e-03 19.38432216 Conventional.2.LL.3
#> 36.102 Conventional       2  LL.3 3.654383e-03 19.34798140 Conventional.2.LL.3
#> 36.103 Conventional       2  LL.3 3.872039e-03 19.30958360 Conventional.2.LL.3
#> 36.104 Conventional       2  LL.3 4.102658e-03 19.26902150 Conventional.2.LL.3
#> 36.105 Conventional       2  LL.3 4.347013e-03 19.22618335 Conventional.2.LL.3
#> 36.106 Conventional       2  LL.3 4.605922e-03 19.18095287 Conventional.2.LL.3
#> 36.107 Conventional       2  LL.3 4.880252e-03 19.13320918 Conventional.2.LL.3
#> 36.108 Conventional       2  LL.3 5.170920e-03 19.08282680 Conventional.2.LL.3
#> 36.109 Conventional       2  LL.3 5.478901e-03 19.02967570 Conventional.2.LL.3
#> 36.110 Conventional       2  LL.3 5.805226e-03 18.97362129 Conventional.2.LL.3
#> 36.111 Conventional       2  LL.3 6.150986e-03 18.91452449 Conventional.2.LL.3
#> 36.112 Conventional       2  LL.3 6.517340e-03 18.85224189 Conventional.2.LL.3
#> 36.113 Conventional       2  LL.3 6.905514e-03 18.78662583 Conventional.2.LL.3
#> 36.114 Conventional       2  LL.3 7.316807e-03 18.71752462 Conventional.2.LL.3
#> 36.115 Conventional       2  LL.3 7.752597e-03 18.64478279 Conventional.2.LL.3
#> 36.116 Conventional       2  LL.3 8.214344e-03 18.56824134 Conventional.2.LL.3
#> 36.117 Conventional       2  LL.3 8.703591e-03 18.48773814 Conventional.2.LL.3
#> 36.118 Conventional       2  LL.3 9.221979e-03 18.40310827 Conventional.2.LL.3
#> 36.119 Conventional       2  LL.3 9.771242e-03 18.31418458 Conventional.2.LL.3
#> 36.120 Conventional       2  LL.3 1.035322e-02 18.22079815 Conventional.2.LL.3
#> 36.121 Conventional       2  LL.3 1.096986e-02 18.12277897 Conventional.2.LL.3
#> 36.122 Conventional       2  LL.3 1.162322e-02 18.01995664 Conventional.2.LL.3
#> 36.123 Conventional       2  LL.3 1.231551e-02 17.91216113 Conventional.2.LL.3
#> 36.124 Conventional       2  LL.3 1.304902e-02 17.79922365 Conventional.2.LL.3
#> 36.125 Conventional       2  LL.3 1.382622e-02 17.68097766 Conventional.2.LL.3
#> 36.126 Conventional       2  LL.3 1.464971e-02 17.55725987 Conventional.2.LL.3
#> 36.127 Conventional       2  LL.3 1.552225e-02 17.42791141 Conventional.2.LL.3
#> 36.128 Conventional       2  LL.3 1.644676e-02 17.29277904 Conventional.2.LL.3
#> 36.129 Conventional       2  LL.3 1.742633e-02 17.15171650 Conventional.2.LL.3
#> 36.130 Conventional       2  LL.3 1.846425e-02 17.00458589 Conventional.2.LL.3
#> 36.131 Conventional       2  LL.3 1.956398e-02 16.85125917 Conventional.2.LL.3
#> 36.132 Conventional       2  LL.3 2.072922e-02 16.69161968 Conventional.2.LL.3
#> 36.133 Conventional       2  LL.3 2.196385e-02 16.52556377 Conventional.2.LL.3
#> 36.134 Conventional       2  LL.3 2.327202e-02 16.35300248 Conventional.2.LL.3
#> 36.135 Conventional       2  LL.3 2.465811e-02 16.17386315 Conventional.2.LL.3
#> 36.136 Conventional       2  LL.3 2.612675e-02 15.98809121 Conventional.2.LL.3
#> 36.137 Conventional       2  LL.3 2.768287e-02 15.79565183 Conventional.2.LL.3
#> 36.138 Conventional       2  LL.3 2.933166e-02 15.59653158 Conventional.2.LL.3
#> 36.139 Conventional       2  LL.3 3.107866e-02 15.39074008 Conventional.2.LL.3
#> 36.140 Conventional       2  LL.3 3.292971e-02 15.17831151 Conventional.2.LL.3
#> 36.141 Conventional       2  LL.3 3.489101e-02 14.95930606 Conventional.2.LL.3
#> 36.142 Conventional       2  LL.3 3.696913e-02 14.73381121 Conventional.2.LL.3
#> 36.143 Conventional       2  LL.3 3.917101e-02 14.50194289 Conventional.2.LL.3
#> 36.144 Conventional       2  LL.3 4.150405e-02 14.26384637 Conventional.2.LL.3
#> 36.145 Conventional       2  LL.3 4.397604e-02 14.01969704 Conventional.2.LL.3
#> 36.146 Conventional       2  LL.3 4.659526e-02 13.76970082 Conventional.2.LL.3
#> 36.147 Conventional       2  LL.3 4.937048e-02 13.51409436 Conventional.2.LL.3
#> 36.148 Conventional       2  LL.3 5.231099e-02 13.25314495 Conventional.2.LL.3
#> 36.149 Conventional       2  LL.3 5.542665e-02 12.98715006 Conventional.2.LL.3
#> 36.150 Conventional       2  LL.3 5.872787e-02 12.71643661 Conventional.2.LL.3
#> 36.151 Conventional       2  LL.3 6.222571e-02 12.44135986 Conventional.2.LL.3
#> 36.152 Conventional       2  LL.3 6.593188e-02 12.16230198 Conventional.2.LL.3
#> 36.153 Conventional       2  LL.3 6.985880e-02 11.87967027 Conventional.2.LL.3
#> 36.154 Conventional       2  LL.3 7.401960e-02 11.59389507 Conventional.2.LL.3
#> 36.155 Conventional       2  LL.3 7.842822e-02 11.30542738 Conventional.2.LL.3
#> 36.156 Conventional       2  LL.3 8.309942e-02 11.01473612 Conventional.2.LL.3
#> 36.157 Conventional       2  LL.3 8.804884e-02 10.72230525 Conventional.2.LL.3
#> 36.158 Conventional       2  LL.3 9.329304e-02 10.42863059 Conventional.2.LL.3
#> 36.159 Conventional       2  LL.3 9.884959e-02 10.13421654 Conventional.2.LL.3
#> 36.160 Conventional       2  LL.3 1.047371e-01  9.83957258 Conventional.2.LL.3
#> 36.161 Conventional       2  LL.3 1.109752e-01  9.54520980 Conventional.2.LL.3
#> 36.162 Conventional       2  LL.3 1.175850e-01  9.25163733 Conventional.2.LL.3
#> 36.163 Conventional       2  LL.3 1.245883e-01  8.95935886 Conventional.2.LL.3
#> 36.164 Conventional       2  LL.3 1.320088e-01  8.66886915 Conventional.2.LL.3
#> 36.165 Conventional       2  LL.3 1.398713e-01  8.38065081 Conventional.2.LL.3
#> 36.166 Conventional       2  LL.3 1.482021e-01  8.09517110 Conventional.2.LL.3
#> 36.167 Conventional       2  LL.3 1.570290e-01  7.81287903 Conventional.2.LL.3
#> 36.168 Conventional       2  LL.3 1.663817e-01  7.53420276 Conventional.2.LL.3
#> 36.169 Conventional       2  LL.3 1.762914e-01  7.25954714 Conventional.2.LL.3
#> 36.170 Conventional       2  LL.3 1.867914e-01  6.98929172 Conventional.2.LL.3
#> 36.171 Conventional       2  LL.3 1.979167e-01  6.72378898 Conventional.2.LL.3
#> 36.172 Conventional       2  LL.3 2.097046e-01  6.46336291 Conventional.2.LL.3
#> 36.173 Conventional       2  LL.3 2.221947e-01  6.20830801 Conventional.2.LL.3
#> 36.174 Conventional       2  LL.3 2.354286e-01  5.95888847 Conventional.2.LL.3
#> 36.175 Conventional       2  LL.3 2.494508e-01  5.71533789 Conventional.2.LL.3
#> 36.176 Conventional       2  LL.3 2.643081e-01  5.47785910 Conventional.2.LL.3
#> 36.177 Conventional       2  LL.3 2.800504e-01  5.24662443 Conventional.2.LL.3
#> 36.178 Conventional       2  LL.3 2.967302e-01  5.02177618 Conventional.2.LL.3
#> 36.179 Conventional       2  LL.3 3.144035e-01  4.80342737 Conventional.2.LL.3
#> 36.180 Conventional       2  LL.3 3.331295e-01  4.59166271 Conventional.2.LL.3
#> 36.181 Conventional       2  LL.3 3.529707e-01  4.38653969 Conventional.2.LL.3
#> 36.182 Conventional       2  LL.3 3.739937e-01  4.18808997 Conventional.2.LL.3
#> 36.183 Conventional       2  LL.3 3.962689e-01  3.99632079 Conventional.2.LL.3
#> 36.184 Conventional       2  LL.3 4.198707e-01  3.81121650 Conventional.2.LL.3
#> 36.185 Conventional       2  LL.3 4.448783e-01  3.63274020 Conventional.2.LL.3
#> 36.186 Conventional       2  LL.3 4.713753e-01  3.46083543 Conventional.2.LL.3
#> 36.187 Conventional       2  LL.3 4.994505e-01  3.29542783 Conventional.2.LL.3
#> 36.188 Conventional       2  LL.3 5.291979e-01  3.13642687 Conventional.2.LL.3
#> 36.189 Conventional       2  LL.3 5.607170e-01  2.98372754 Conventional.2.LL.3
#> 36.190 Conventional       2  LL.3 5.941134e-01  2.83721195 Conventional.2.LL.3
#> 36.191 Conventional       2  LL.3 6.294989e-01  2.69675101 Conventional.2.LL.3
#> 36.192 Conventional       2  LL.3 6.669920e-01  2.56220591 Conventional.2.LL.3
#> 36.193 Conventional       2  LL.3 7.067181e-01  2.43342963 Conventional.2.LL.3
#> 36.194 Conventional       2  LL.3 7.488104e-01  2.31026832 Conventional.2.LL.3
#> 36.195 Conventional       2  LL.3 7.934097e-01  2.19256259 Conventional.2.LL.3
#> 36.196 Conventional       2  LL.3 8.406653e-01  2.08014879 Conventional.2.LL.3
#> 36.197 Conventional       2  LL.3 8.907355e-01  1.97286009 Conventional.2.LL.3
#> 36.198 Conventional       2  LL.3 9.437878e-01  1.87052755 Conventional.2.LL.3
#> 36.199 Conventional       2  LL.3 1.000000e+00  1.77298107 Conventional.2.LL.3
#> 71          Organic       3  LL.3 1.000000e-05 20.04010725      Organic.3.LL.3
#> 71.1        Organic       3  LL.3 1.059560e-05 20.03856324      Organic.3.LL.3
#> 71.2        Organic       3  LL.3 1.122668e-05 20.03691386      Organic.3.LL.3
#> 71.3        Organic       3  LL.3 1.189534e-05 20.03515193      Organic.3.LL.3
#> 71.4        Organic       3  LL.3 1.260383e-05 20.03326980      Organic.3.LL.3
#> 71.5        Organic       3  LL.3 1.335452e-05 20.03125929      Organic.3.LL.3
#> 71.6        Organic       3  LL.3 1.414991e-05 20.02911168      Organic.3.LL.3
#> 71.7        Organic       3  LL.3 1.499268e-05 20.02681766      Organic.3.LL.3
#> 71.8        Organic       3  LL.3 1.588565e-05 20.02436726      Organic.3.LL.3
#> 71.9        Organic       3  LL.3 1.683180e-05 20.02174988      Organic.3.LL.3
#> 71.10       Organic       3  LL.3 1.783431e-05 20.01895420      Organic.3.LL.3
#> 71.11       Organic       3  LL.3 1.889652e-05 20.01596810      Organic.3.LL.3
#> 71.12       Organic       3  LL.3 2.002200e-05 20.01277870      Organic.3.LL.3
#> 71.13       Organic       3  LL.3 2.121452e-05 20.00937221      Organic.3.LL.3
#> 71.14       Organic       3  LL.3 2.247806e-05 20.00573395      Organic.3.LL.3
#> 71.15       Organic       3  LL.3 2.381686e-05 20.00184824      Organic.3.LL.3
#> 71.16       Organic       3  LL.3 2.523539e-05 19.99769835      Organic.3.LL.3
#> 71.17       Organic       3  LL.3 2.673842e-05 19.99326645      Organic.3.LL.3
#> 71.18       Organic       3  LL.3 2.833096e-05 19.98853351      Organic.3.LL.3
#> 71.19       Organic       3  LL.3 3.001836e-05 19.98347924      Organic.3.LL.3
#> 71.20       Organic       3  LL.3 3.180626e-05 19.97808200      Organic.3.LL.3
#> 71.21       Organic       3  LL.3 3.370064e-05 19.97231874      Organic.3.LL.3
#> 71.22       Organic       3  LL.3 3.570786e-05 19.96616484      Organic.3.LL.3
#> 71.23       Organic       3  LL.3 3.783463e-05 19.95959412      Organic.3.LL.3
#> 71.24       Organic       3  LL.3 4.008806e-05 19.95257864      Organic.3.LL.3
#> 71.25       Organic       3  LL.3 4.247572e-05 19.94508864      Organic.3.LL.3
#> 71.26       Organic       3  LL.3 4.500558e-05 19.93709241      Organic.3.LL.3
#> 71.27       Organic       3  LL.3 4.768612e-05 19.92855621      Organic.3.LL.3
#> 71.28       Organic       3  LL.3 5.052631e-05 19.91944407      Organic.3.LL.3
#> 71.29       Organic       3  LL.3 5.353567e-05 19.90971773      Organic.3.LL.3
#> 71.30       Organic       3  LL.3 5.672426e-05 19.89933644      Organic.3.LL.3
#> 71.31       Organic       3  LL.3 6.010277e-05 19.88825687      Organic.3.LL.3
#> 71.32       Organic       3  LL.3 6.368250e-05 19.87643291      Organic.3.LL.3
#> 71.33       Organic       3  LL.3 6.747544e-05 19.86381554      Organic.3.LL.3
#> 71.34       Organic       3  LL.3 7.149429e-05 19.85035264      Organic.3.LL.3
#> 71.35       Organic       3  LL.3 7.575250e-05 19.83598882      Organic.3.LL.3
#> 71.36       Organic       3  LL.3 8.026434e-05 19.82066524      Organic.3.LL.3
#> 71.37       Organic       3  LL.3 8.504489e-05 19.80431944      Organic.3.LL.3
#> 71.38       Organic       3  LL.3 9.011018e-05 19.78688509      Organic.3.LL.3
#> 71.39       Organic       3  LL.3 9.547716e-05 19.76829186      Organic.3.LL.3
#> 71.40       Organic       3  LL.3 1.011638e-04 19.74846514      Organic.3.LL.3
#> 71.41       Organic       3  LL.3 1.071891e-04 19.72732586      Organic.3.LL.3
#> 71.42       Organic       3  LL.3 1.135733e-04 19.70479028      Organic.3.LL.3
#> 71.43       Organic       3  LL.3 1.203378e-04 19.68076975      Organic.3.LL.3
#> 71.44       Organic       3  LL.3 1.275051e-04 19.65517047      Organic.3.LL.3
#> 71.45       Organic       3  LL.3 1.350994e-04 19.62789329      Organic.3.LL.3
#> 71.46       Organic       3  LL.3 1.431459e-04 19.59883347      Organic.3.LL.3
#> 71.47       Organic       3  LL.3 1.516717e-04 19.56788045      Organic.3.LL.3
#> 71.48       Organic       3  LL.3 1.607053e-04 19.53491761      Organic.3.LL.3
#> 71.49       Organic       3  LL.3 1.702769e-04 19.49982210      Organic.3.LL.3
#> 71.50       Organic       3  LL.3 1.804186e-04 19.46246459      Organic.3.LL.3
#> 71.51       Organic       3  LL.3 1.911644e-04 19.42270910      Organic.3.LL.3
#> 71.52       Organic       3  LL.3 2.025502e-04 19.38041279      Organic.3.LL.3
#> 71.53       Organic       3  LL.3 2.146141e-04 19.33542585      Organic.3.LL.3
#> 71.54       Organic       3  LL.3 2.273966e-04 19.28759131      Organic.3.LL.3
#> 71.55       Organic       3  LL.3 2.409404e-04 19.23674500      Organic.3.LL.3
#> 71.56       Organic       3  LL.3 2.552908e-04 19.18271541      Organic.3.LL.3
#> 71.57       Organic       3  LL.3 2.704960e-04 19.12532372      Organic.3.LL.3
#> 71.58       Organic       3  LL.3 2.866068e-04 19.06438381      Organic.3.LL.3
#> 71.59       Organic       3  LL.3 3.036771e-04 18.99970235      Organic.3.LL.3
#> 71.60       Organic       3  LL.3 3.217642e-04 18.93107889      Organic.3.LL.3
#> 71.61       Organic       3  LL.3 3.409285e-04 18.85830618      Organic.3.LL.3
#> 71.62       Organic       3  LL.3 3.612343e-04 18.78117039      Organic.3.LL.3
#> 71.63       Organic       3  LL.3 3.827494e-04 18.69945155      Organic.3.LL.3
#> 71.64       Organic       3  LL.3 4.055461e-04 18.61292403      Organic.3.LL.3
#> 71.65       Organic       3  LL.3 4.297005e-04 18.52135717      Organic.3.LL.3
#> 71.66       Organic       3  LL.3 4.552935e-04 18.42451604      Organic.3.LL.3
#> 71.67       Organic       3  LL.3 4.824109e-04 18.32216225      Organic.3.LL.3
#> 71.68       Organic       3  LL.3 5.111433e-04 18.21405507      Organic.3.LL.3
#> 71.69       Organic       3  LL.3 5.415871e-04 18.09995252      Organic.3.LL.3
#> 71.70       Organic       3  LL.3 5.738442e-04 17.97961275      Organic.3.LL.3
#> 71.71       Organic       3  LL.3 6.080224e-04 17.85279559      Organic.3.LL.3
#> 71.72       Organic       3  LL.3 6.442364e-04 17.71926416      Organic.3.LL.3
#> 71.73       Organic       3  LL.3 6.826072e-04 17.57878683      Organic.3.LL.3
#> 71.74       Organic       3  LL.3 7.232634e-04 17.43113921      Organic.3.LL.3
#> 71.75       Organic       3  LL.3 7.663411e-04 17.27610644      Organic.3.LL.3
#> 71.76       Organic       3  LL.3 8.119845e-04 17.11348556      Organic.3.LL.3
#> 71.77       Organic       3  LL.3 8.603464e-04 16.94308810      Organic.3.LL.3
#> 71.78       Organic       3  LL.3 9.115888e-04 16.76474282      Organic.3.LL.3
#> 71.79       Organic       3  LL.3 9.658832e-04 16.57829852      Organic.3.LL.3
#> 71.80       Organic       3  LL.3 1.023411e-03 16.38362702      Organic.3.LL.3
#> 71.81       Organic       3  LL.3 1.084366e-03 16.18062618      Organic.3.LL.3
#> 71.82       Organic       3  LL.3 1.148951e-03 15.96922293      Organic.3.LL.3
#> 71.83       Organic       3  LL.3 1.217383e-03 15.74937631      Organic.3.LL.3
#> 71.84       Organic       3  LL.3 1.289890e-03 15.52108044      Organic.3.LL.3
#> 71.85       Organic       3  LL.3 1.366716e-03 15.28436735      Organic.3.LL.3
#> 71.86       Organic       3  LL.3 1.448118e-03 15.03930967      Organic.3.LL.3
#> 71.87       Organic       3  LL.3 1.534368e-03 14.78602301      Organic.3.LL.3
#> 71.88       Organic       3  LL.3 1.625756e-03 14.52466806      Organic.3.LL.3
#> 71.89       Organic       3  LL.3 1.722586e-03 14.25545231      Organic.3.LL.3
#> 71.90       Organic       3  LL.3 1.825183e-03 13.97863134      Organic.3.LL.3
#> 71.91       Organic       3  LL.3 1.933892e-03 13.69450953      Organic.3.LL.3
#> 71.92       Organic       3  LL.3 2.049075e-03 13.40344030      Organic.3.LL.3
#> 71.93       Organic       3  LL.3 2.171118e-03 13.10582565      Organic.3.LL.3
#> 71.94       Organic       3  LL.3 2.300430e-03 12.80211512      Organic.3.LL.3
#> 71.95       Organic       3  LL.3 2.437444e-03 12.49280404      Organic.3.LL.3
#> 71.96       Organic       3  LL.3 2.582619e-03 12.17843107      Organic.3.LL.3
#> 71.97       Organic       3  LL.3 2.736440e-03 11.85957514      Organic.3.LL.3
#> 71.98       Organic       3  LL.3 2.899423e-03 11.53685166      Organic.3.LL.3
#> 71.99       Organic       3  LL.3 3.072113e-03 11.21090816      Organic.3.LL.3
#> 71.100      Organic       3  LL.3 3.255089e-03 10.88241936      Organic.3.LL.3
#> 71.101      Organic       3  LL.3 3.448962e-03 10.55208175      Organic.3.LL.3
#> 71.102      Organic       3  LL.3 3.654383e-03 10.22060782      Organic.3.LL.3
#> 71.103      Organic       3  LL.3 3.872039e-03  9.88871992      Organic.3.LL.3
#> 71.104      Organic       3  LL.3 4.102658e-03  9.55714405      Organic.3.LL.3
#> 71.105      Organic       3  LL.3 4.347013e-03  9.22660346      Organic.3.LL.3
#> 71.106      Organic       3  LL.3 4.605922e-03  8.89781240      Organic.3.LL.3
#> 71.107      Organic       3  LL.3 4.880252e-03  8.57146996      Organic.3.LL.3
#> 71.108      Organic       3  LL.3 5.170920e-03  8.24825426      Organic.3.LL.3
#> 71.109      Organic       3  LL.3 5.478901e-03  7.92881696      Organic.3.LL.3
#> 71.110      Organic       3  LL.3 5.805226e-03  7.61377825      Organic.3.LL.3
#> 71.111      Organic       3  LL.3 6.150986e-03  7.30372244      Organic.3.LL.3
#> 71.112      Organic       3  LL.3 6.517340e-03  6.99919405      Organic.3.LL.3
#> 71.113      Organic       3  LL.3 6.905514e-03  6.70069465      Organic.3.LL.3
#> 71.114      Organic       3  LL.3 7.316807e-03  6.40868035      Organic.3.LL.3
#> 71.115      Organic       3  LL.3 7.752597e-03  6.12355990      Organic.3.LL.3
#> 71.116      Organic       3  LL.3 8.214344e-03  5.84569358      Organic.3.LL.3
#> 71.117      Organic       3  LL.3 8.703591e-03  5.57539268      Organic.3.LL.3
#> 71.118      Organic       3  LL.3 9.221979e-03  5.31291959      Organic.3.LL.3
#> 71.119      Organic       3  LL.3 9.771242e-03  5.05848850      Organic.3.LL.3
#> 71.120      Organic       3  LL.3 1.035322e-02  4.81226657      Organic.3.LL.3
#> 71.121      Organic       3  LL.3 1.096986e-02  4.57437565      Organic.3.LL.3
#> 71.122      Organic       3  LL.3 1.162322e-02  4.34489424      Organic.3.LL.3
#> 71.123      Organic       3  LL.3 1.231551e-02  4.12385993      Organic.3.LL.3
#> 71.124      Organic       3  LL.3 1.304902e-02  3.91127201      Organic.3.LL.3
#> 71.125      Organic       3  LL.3 1.382622e-02  3.70709426      Organic.3.LL.3
#> 71.126      Organic       3  LL.3 1.464971e-02  3.51125794      Organic.3.LL.3
#> 71.127      Organic       3  LL.3 1.552225e-02  3.32366482      Organic.3.LL.3
#> 71.128      Organic       3  LL.3 1.644676e-02  3.14419018      Organic.3.LL.3
#> 71.129      Organic       3  LL.3 1.742633e-02  2.97268589      Organic.3.LL.3
#> 71.130      Organic       3  LL.3 1.846425e-02  2.80898334      Organic.3.LL.3
#> 71.131      Organic       3  LL.3 1.956398e-02  2.65289632      Organic.3.LL.3
#> 71.132      Organic       3  LL.3 2.072922e-02  2.50422374      Organic.3.LL.3
#> 71.133      Organic       3  LL.3 2.196385e-02  2.36275227      Organic.3.LL.3
#> 71.134      Organic       3  LL.3 2.327202e-02  2.22825872      Organic.3.LL.3
#> 71.135      Organic       3  LL.3 2.465811e-02  2.10051235      Organic.3.LL.3
#> 71.136      Organic       3  LL.3 2.612675e-02  1.97927689      Organic.3.LL.3
#> 71.137      Organic       3  LL.3 2.768287e-02  1.86431254      Organic.3.LL.3
#> 71.138      Organic       3  LL.3 2.933166e-02  1.75537757      Organic.3.LL.3
#> 71.139      Organic       3  LL.3 3.107866e-02  1.65222996      Organic.3.LL.3
#> 71.140      Organic       3  LL.3 3.292971e-02  1.55462870      Organic.3.LL.3
#> 71.141      Organic       3  LL.3 3.489101e-02  1.46233503      Organic.3.LL.3
#> 71.142      Organic       3  LL.3 3.696913e-02  1.37511347      Organic.3.LL.3
#> 71.143      Organic       3  LL.3 3.917101e-02  1.29273271      Organic.3.LL.3
#> 71.144      Organic       3  LL.3 4.150405e-02  1.21496638      Organic.3.LL.3
#> 71.145      Organic       3  LL.3 4.397604e-02  1.14159367      Organic.3.LL.3
#> 71.146      Organic       3  LL.3 4.659526e-02  1.07239988      Organic.3.LL.3
#> 71.147      Organic       3  LL.3 4.937048e-02  1.00717679      Organic.3.LL.3
#> 71.148      Organic       3  LL.3 5.231099e-02  0.94572300      Organic.3.LL.3
#> 71.149      Organic       3  LL.3 5.542665e-02  0.88784416      Organic.3.LL.3
#> 71.150      Organic       3  LL.3 5.872787e-02  0.83335313      Organic.3.LL.3
#> 71.151      Organic       3  LL.3 6.222571e-02  0.78207005      Organic.3.LL.3
#> 71.152      Organic       3  LL.3 6.593188e-02  0.73382242      Organic.3.LL.3
#> 71.153      Organic       3  LL.3 6.985880e-02  0.68844500      Organic.3.LL.3
#> 71.154      Organic       3  LL.3 7.401960e-02  0.64577985      Organic.3.LL.3
#> 71.155      Organic       3  LL.3 7.842822e-02  0.60567613      Organic.3.LL.3
#> 71.156      Organic       3  LL.3 8.309942e-02  0.56799005      Organic.3.LL.3
#> 71.157      Organic       3  LL.3 8.804884e-02  0.53258466      Organic.3.LL.3
#> 71.158      Organic       3  LL.3 9.329304e-02  0.49932973      Organic.3.LL.3
#> 71.159      Organic       3  LL.3 9.884959e-02  0.46810149      Organic.3.LL.3
#> 71.160      Organic       3  LL.3 1.047371e-01  0.43878247      Organic.3.LL.3
#> 71.161      Organic       3  LL.3 1.109752e-01  0.41126127      Organic.3.LL.3
#> 71.162      Organic       3  LL.3 1.175850e-01  0.38543234      Organic.3.LL.3
#> 71.163      Organic       3  LL.3 1.245883e-01  0.36119576      Organic.3.LL.3
#> 71.164      Organic       3  LL.3 1.320088e-01  0.33845700      Organic.3.LL.3
#> 71.165      Organic       3  LL.3 1.398713e-01  0.31712670      Organic.3.LL.3
#> 71.166      Organic       3  LL.3 1.482021e-01  0.29712043      Organic.3.LL.3
#> 71.167      Organic       3  LL.3 1.570290e-01  0.27835849      Organic.3.LL.3
#> 71.168      Organic       3  LL.3 1.663817e-01  0.26076565      Organic.3.LL.3
#> 71.169      Organic       3  LL.3 1.762914e-01  0.24427099      Organic.3.LL.3
#> 71.170      Organic       3  LL.3 1.867914e-01  0.22880764      Organic.3.LL.3
#> 71.171      Organic       3  LL.3 1.979167e-01  0.21431260      Organic.3.LL.3
#> 71.172      Organic       3  LL.3 2.097046e-01  0.20072653      Organic.3.LL.3
#> 71.173      Organic       3  LL.3 2.221947e-01  0.18799357      Organic.3.LL.3
#> 71.174      Organic       3  LL.3 2.354286e-01  0.17606116      Organic.3.LL.3
#> 71.175      Organic       3  LL.3 2.494508e-01  0.16487985      Organic.3.LL.3
#> 71.176      Organic       3  LL.3 2.643081e-01  0.15440313      Organic.3.LL.3
#> 71.177      Organic       3  LL.3 2.800504e-01  0.14458728      Organic.3.LL.3
#> 71.178      Organic       3  LL.3 2.967302e-01  0.13539120      Organic.3.LL.3
#> 71.179      Organic       3  LL.3 3.144035e-01  0.12677629      Organic.3.LL.3
#> 71.180      Organic       3  LL.3 3.331295e-01  0.11870629      Organic.3.LL.3
#> 71.181      Organic       3  LL.3 3.529707e-01  0.11114712      Organic.3.LL.3
#> 71.182      Organic       3  LL.3 3.739937e-01  0.10406680      Organic.3.LL.3
#> 71.183      Organic       3  LL.3 3.962689e-01  0.09743531      Organic.3.LL.3
#> 71.184      Organic       3  LL.3 4.198707e-01  0.09122447      Organic.3.LL.3
#> 71.185      Organic       3  LL.3 4.448783e-01  0.08540784      Organic.3.LL.3
#> 71.186      Organic       3  LL.3 4.713753e-01  0.07996060      Organic.3.LL.3
#> 71.187      Organic       3  LL.3 4.994505e-01  0.07485948      Organic.3.LL.3
#> 71.188      Organic       3  LL.3 5.291979e-01  0.07008264      Organic.3.LL.3
#> 71.189      Organic       3  LL.3 5.607170e-01  0.06560962      Organic.3.LL.3
#> 71.190      Organic       3  LL.3 5.941134e-01  0.06142121      Organic.3.LL.3
#> 71.191      Organic       3  LL.3 6.294989e-01  0.05749942      Organic.3.LL.3
#> 71.192      Organic       3  LL.3 6.669920e-01  0.05382735      Organic.3.LL.3
#> 71.193      Organic       3  LL.3 7.067181e-01  0.05038921      Organic.3.LL.3
#> 71.194      Organic       3  LL.3 7.488104e-01  0.04717016      Organic.3.LL.3
#> 71.195      Organic       3  LL.3 7.934097e-01  0.04415629      Organic.3.LL.3
#> 71.196      Organic       3  LL.3 8.406653e-01  0.04133460      Organic.3.LL.3
#> 71.197      Organic       3  LL.3 8.907355e-01  0.03869287      Organic.3.LL.3
#> 71.198      Organic       3  LL.3 9.437878e-01  0.03621967      Organic.3.LL.3
#> 71.199      Organic       3  LL.3 1.000000e+00  0.03390428      Organic.3.LL.3