Estimate Efective dose (EC50) for multi isolate data set using one or more models

Estimate Efective dose (EC50) for multi isolate stratified data set using one or more models.

ec50_multimodel(formula, data, EC_lvl = 50,
isolate_col, strata_col= NULL,
fct, interval = c("none", "delta", "fls", "tfls"),
type = c("relative", "absolute"))

Arguments

formula

An object of class "formula.(e.g. growth ~ dose")

data

A data.frame in which the dose-response data is in

EC_lvl

Define the EC level. Default is 50

isolate_col

indicate the isolate colunm. Indicate the name inside "". (e.g. "isolates")

strata_col

indicate the strata colunms. If there are more then one colunms, indicate as a vector (e.g. c("region", "field"))

fct

A list specifying the non-linear models to be fitted. Please, inform the model functions with the :: operator to inform the drc package (e.g. list(drc::LL.4(), drc::LL.3())). For more info, see "drc::drm".

interval

A character string specifying the type of confidence intervals to be supplied. For more information see "ED"

type

Whether the specified response levels are absolute or relative (default)

Value

A dataframe containing EC estimates and statistics for model selection for each model used. For more info, see "drc::ED".

Examples

data(multi_isolate)

# Load data
ec50_multimodel(growth ~ dose,
                data = multi_isolate,
                isolate_col = "isolate",
                strata_col = c("fungicida", "field"),
                fct = list(drc::LL.5(),drc::W2.3(), drc::LL2.3()))