Select EC50 candidate models

Rank candidate models fitted with [ec50_multimodel()] within each isolate and stratum using an information criterion such as `IC`.

Usage

model_selection(x, criterion = "IC")

best_model(x, criterion = "IC")

Arguments

x

An object returned by [ec50_multimodel()].

criterion

Character scalar naming the criterion column. Smaller values are considered better.

Value

A data frame with model rankings, delta criterion values, and weights.

Examples

data(multi_isolate)
sample_data <- subset(multi_isolate, isolate %in% 1:3 & fungicida == "Fungicide A")
fit <- ec50_multimodel(
  growth ~ dose,
  data = sample_data,
  isolate_col = "isolate",
  strata_col = "field",
  fct = list(drc::LL.3(), drc::LL.4())
)
model_selection(fit)
#>   ID        field    Estimate   Std..Error    logLik       IC Lack.of.fit
#> 1  2 Conventional 0.101455765 0.0076364691 -43.53183 95.06365   0.9866424
#> 2  2 Conventional 0.098010789 0.0128129346 -43.47914 96.95829   0.9683378
#> 3  1      Organic 0.006072082 0.0005740341 -45.15079 98.30158   0.7271292
#> 4  1      Organic 0.006364103 0.0007031475 -44.69257 99.38515   0.7370811
#> 5  3      Organic 0.003602158 0.0002580833 -34.85654 79.71307   0.5658418
#> 6  3      Organic 0.003776957 0.0002432571 -36.09845 80.19689   0.3897432
#>     Res.var model     delta    weight rank
#> 1 0.7704874  LL.3 0.0000000 0.7205753    1
#> 2 0.7929511  LL.4 1.8946342 0.2794247    2
#> 3 0.8451681  LL.3 0.0000000 0.6322276    1
#> 4 0.8498845  LL.4 1.0835703 0.3677724    2
#> 5 0.4844631  LL.4 0.0000000 0.5601838    1
#> 6 0.5038400  LL.3 0.4838157 0.4398162    2
best_model(fit)
#>   ID        field    Estimate   Std..Error    logLik       IC Lack.of.fit
#> 1  2 Conventional 0.101455765 0.0076364691 -43.53183 95.06365   0.9866424
#> 2  1      Organic 0.006072082 0.0005740341 -45.15079 98.30158   0.7271292
#> 3  3      Organic 0.003602158 0.0002580833 -34.85654 79.71307   0.5658418
#>     Res.var model delta    weight rank
#> 1 0.7704874  LL.3     0 0.7205753    1
#> 2 0.8451681  LL.3     0 0.6322276    1
#> 3 0.4844631  LL.4     0 0.5601838    1