Estimate Efective dose (EC50) for multi isolate data set

Estimate Efective dose (EC50) for multi isolate stratified data set.

estimate_EC50(formula, data, EC_lvl = 50, isolate_col,
strata_col = NULL,
fct,interval = c("none", "delta", "fls", "tfls"))

Arguments

formula
An object of class "formula.(e.g. growth ~ dose")

data
A data.frame in which the dose-response data is in

EC_lvl
Define the EC level. Default is 50

isolate_col
indicate the isolate colunm. Indicate the name inside "". (e.g. "isolates")

strata_col
indicate the strata colunms. If there are more then one colunms, indicate as a vector (e.g. c("region", "field"))

fct
A list with three or more elements specifying the non-linear function. Please, inform the model function with the :: operator to inform the drc package (e.g. drc::LL.4()). For more info, see "drm".

interval
A character string specifying the type of confidence intervals to be supplied. For more information see "ED"

Examples

# Load data
data(multi_isolate)

df_ec50 = estimate_EC50(growth~dose,
                        data =multi_isolate,
                        isolate_col = "isolate",
                        strata_col =  c("field","fungicida"),
                        interval = "delta",
                        fct = drc::LL.3())
head(df_ec50)

#>   ID   field   fungicida    Estimate   Std..Error       Lower       Upper
#> 1  1 Organic Fungicide A 0.006072082 0.0005740341 0.004902813 0.007241351
#> 2  3 Organic Fungicide A 0.003776957 0.0002432571 0.003281459 0.004272456
#> 3  5 Organic Fungicide A 0.006122508 0.0004575060 0.005190599 0.007054418
#> 4  7 Organic Fungicide A 0.006532201 0.0005278730 0.005456959 0.007607443
#> 5  9 Organic Fungicide A 0.005115267 0.0004672111 0.004163589 0.006066945
#> 6 11 Organic Fungicide A 0.005044116 0.0005000569 0.004025533 0.006062698